Fix test_exceptions test and adjust reversal size logic
Previous implementation didn't account for the movement after the last reversal.  This led to issues with both the final box value and the final column volume data.  This change adds logic to track any further movement (volume and box) in the same signed direction after the final reversal and adds it to the last box column after the clean data loop.