Removed `if (plottype not in _valid_types_all) ... `
This is unnecessary because the kwarg validator already check whether `kwarg['type']` is in `_valid_types_all`. So if the value passed the validator then it must be in `_valid_types_all`. So we don't need to check again in function `_get_valid_plot_types` (unless...this function is used also by function other than `plotting.plot`)